// react-component-library/src/core/enums/graphql-translations/graphql-translation-attributes-fields.enum.ts
export enum EGraphqlTranslationAttributesFields {
  title = 'ObjObjectTitleGrp.TitleTxt',
  technicalTerm = 'ObjTechnicalTermClb',
  dating = 'ObjDateGrp.DateTxt',
  identNumber = 'ObjObjectNumberGrp.InventarNrSTxt',
  collectionKey = '__orgUnit',
  location = 'ObjCurrentLocationGrpVrt',
  involvedParties = 'ObjPerAssociationMainParticipantVrt',
  description = 'ObjTextOnlineGrp.TextClb',
  geographicalReferences = 'ObjGeograficGrp.PlaceVoc',
  materialAndTechnique = 'ObjMaterialTechniqueGrp.ExportClb',
  dimensionsAndWeight = 'ObjDimAllGrp.PreviewVrt',
}